requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
django-filter==23.2
celery==5.3.1
pytest==7.4.1
pytest-django==4.5.2
